+++ admin/modules/user/index.php
@@ -4,7 +4,9 @@
 
    require_once __DIR__. "/../../autoload/autoload.php";
 
-   
+   $sqlLevel = "SELECT * FROM admin WHERE id = '".$_SESSION['admin_id']."'";
+
+   $level = $db->fetchsql($sqlLevel);
 
    // $product = $db->fetch_join("product");
    $user = $db->fetchAll("user");
@@ -43,7 +45,11 @@
                                           <th class="sorting" tabindex="0" aria-controls="dataTable" rowspan="1" colspan="1" aria-label="Position: activate to sort column ascending">Họ tên</th>
                                           <th class="sorting" tabindex="0" aria-controls="dataTable" rowspan="1" colspan="1" aria-label="Position: activate to sort column ascending">Số điện thoại</th>
                                           <th class="sorting" tabindex="0" aria-controls="dataTable" rowspan="1" colspan="1" aria-label="Position: activate to sort column ascending">Email</th>
-                                          <th class="sorting" tabindex="0" aria-controls="dataTable" rowspan="1" colspan="1" aria-label="Age: activate to sort column ascending">Action</th>
+                                          <?php foreach ($level as $item): ?>
+                                             <?php if ($item['level'] == 1): ?>
+                                                <th class="sorting" tabindex="0" aria-controls="dataTable" rowspan="1" colspan="1" aria-label="Age: activate to sort column ascending">Action</th>
+                                             <?php endif ?>
+                                          <?php endforeach ?>
                                        </tr>
                                     </thead>
                                     <tbody>
@@ -53,9 +59,13 @@
                                              <td><?php echo $item['name'] ?></td>
                                              <td><?php echo $item['phone'] ?></td>
                                              <td><?php echo $item['email'] ?></td>
-                                             <td>
-                                                <a class="btn btn-xs btn-danger" href="delete.php?id=<?php echo $item['id'] ?>"><i class="fa fa-times"></i> Xóa</a>
-                                             </td>
+                                             <?php foreach ($level as $item): ?>
+                                                <?php if ($item['level'] == 1): ?>
+                                                   <td>
+                                                      <a class="btn btn-xs btn-danger" href="delete.php?id=<?php echo $item['id'] ?>"><i class="fa fa-times"></i> Xóa</a>
+                                                   </td>
+                                                <?php endif ?>
+                                             <?php endforeach ?>
                                           </tr>
                                        <?php $stt++; endforeach ?>
